PWN-5022 - Added resizing for input view

+++ app/src/main/java/org/p2p/wallet/solend/ui/withdraw/SolendWithdrawFragment.kt
@@ -63,7 +63,7 @@ class SolendWithdrawFragment :
             viewDoubleInput.setInputData(
                 inputSymbol = token.tokenSymbol,
                 outputSymbol = USD_READABLE_SYMBOL, // todo: talk to managers about output
-                inputRate = token.usdRate.toFloat()
+                inputRate = token.usdRate.toDouble()
             )
         }
     }
